requirements.txt
# Core dependencies
yt-dlp
faster-whisper
torchaudio
torchaudio
transformers

# Speaker diarization
pyannote.audio

# Audio processing
librosa
pydub
soundfile

# Utilities
langdetect
numpy
python-dotenv